desktop/requirements.txt
PyQt5==5.15.9
matplotlib==3.8.2
requests==2.31.0
python-dotenv==1.0.0
